requirements.txt
accelerate==0.31.0
annotated-types==0.7.0
ansible==10.1.0
ansible-core==2.17.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
ascii-magic==2.3.0
asttokens==2.2.1
astunparse==1.6.3
asyncache==0.3.1
atomicwrites==1.4.0
azure-core==1.30.2
azure-storage-blob==12.19.1
backcall==0.2.0
backgroundremover==0.2.8
backports.weakref==1.0.post1
bcrypt==4.1.2
bitsandbytes==0.42.0
bleach==6.0.0
boto3==1.34.56
botocore==1.34.129
cachetools==5.3.3
clyent==1.2.2
commandlines==0.4.1
conda-repo-cli==1.0.75
conda-verify==3.4.2
contourpy==1.1.0
cryptography==41.0.7
cv==1.0.0
Cython==0.29.37
dataclasses-json==0.6.6
decorator==4.4.2
distlib==0.3.8
dnspython==2.6.1
docker==6.1.3
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
email_validator==2.1.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
faiss-cpu==1.8.0
Farama-Notifications==0.0.4
fastapi==0.110.0
fastjsonschema==2.18.0
ffmpeg-python==0.2.0
filelock==3.12.2
filetype==1.2.0
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.40.0
forbiddenfruit==0.1.4
gevent==23.9.1
google-api-core==2.19.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.14.0
h5py==3.11.0
hagrid==0.3.124
hnswlib==0.7.0
hsh==1.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
imageio==2.31.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
ipykernel==5.5.5
ipython==8.12.3
isodate==0.6.1
itables==1.7.1
jax==0.4.20
jax-jumpy==1.0.0
jaxlib==0.4.20
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.0
jupyter-console==6.4.4
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keras==3.3.3
kornia==0.7.2
kornia_rs==0.1.3
kr8s==0.13.5
kviz==0.0.7
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.1
langsmith==0.1.75
lazy_loader==0.2
libclang==18.1.1
loguru==0.7.2
marshmallow==3.21.3
matplotlib==3.8.3
mediapy @ git+https://github.com/PWhiddy/mediapy.git@45101800d4f6adeffe814cad93de1db67c1bd614
mistune==3.0.1
ml-dtypes==0.3.2
moviepy==1.0.3
munkres==1.1.4
names==0.3.0
namex==0.0.8
nbclassic==1.0.0
nbconvert==7.16.4
networkx==3.2.1
nnv==0.0.5
notebook==6.5.6
numpy==1.25.0
openai==1.31.1
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
openpyxl==3.0.10
optree==0.11.0
orjson==3.10.3
packaging==23.2
pandas==2.2.1
paramiko==3.4.0
pathtools==0.1.2
patsy==0.5.3
piexif==1.1.3
Pillow==9.5.0
pipreqs==0.5.0
ply==3.11
proglog==0.1.10
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
pyarrow==15.0.0
pyasn1-modules==0.2.8
pyboy==1.6.9
pycapnp==2.0.0b2
pycurl==7.45.2
pydantic==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.1
pydicom==2.4.4
PyDispatcher==2.0.5
pygraphviz==1.13
pylibjpeg==2.0.0
pylibjpeg-libjpeg==2.2.0
pylibjpeg-openjpeg==2.2.1
pyls-spyder==0.4.0
PyMatting==1.1.12
pymongo==4.6.2
PyNaCl==1.5.0
pyparsing==3.1.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
PySDL2==0.9.15
pysdl2-dll==2.28.0
python-box==7.2.0
python-jsonpath==1.1.1
python-lsp-jsonrpc==1.0.0
pytz==2023.3
PyWavelets==1.4.1
pyzmq==25.1.1
requests-oauthlib==1.3.0
resolvelib==1.0.1
RestrictedPython==7.0
result==0.16.1
rpds-py==0.9.2
rsa==4.9
s3transfer==0.10.1
scikit-image==0.21.0
scipy==1.10.1
segment-anything==1.0
sentence-transformers==3.0.1
sentry-sdk==2.3.1
setproctitle==1.3.3
sherlock==0.4.1
Shimmy==1.1.0
soupsieve==2.4.1
spandrel==0.3.4
stable-baselines3==2.0.0
stack-data==0.6.2
starlette==0.36.3
syft==0.8.6
tensorboard==2.17.1
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.0
tf_keras==2.17.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchsde==0.2.6
tqdm==4.66.2
traitlets==5.9.0
trampoline==0.1.2
transformers==4.41.2
typeguard==4.1.5
typing-inspect==0.9.0
typing_extensions==4.10.0
ultralytics==8.2.27
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.26.2
virtualenv-api==2.1.18
waitress==3.0.0
wandb==0.15.12
watchfiles==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
zope.event==5.0
